pyproject.toml
[project]
name = "snapcook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.102.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.48.1",
    "ultralytics>=8.3.186",
]
